rails 4.1/4.2 specs
