+++ lib/active_record/connection_adapters/rails_42.rb
@@ -1,369 +1,7 @@
-require 'active_record/connection_adapters/abstract_adapter'
-require 'active_record/connection_adapters/statement_pool'
-require 'active_record/connection_adapters/abstract/schema_statements'
-require 'arel/visitors/bind_visitor'
+
 
 module ActiveRecord
 
-  module BigQueryRailsHelpers
-    def self.rails40?
-      ActiveRecord::VERSION::MAJOR == 4 && ActiveRecord::VERSION::MINOR == 0
-    end
-
-    def self.rails41?
-      ActiveRecord::VERSION::MAJOR == 4 && ActiveRecord::VERSION::MINOR == 1
-    end
-
-    def self.rails42?
-      ActiveRecord::VERSION::MAJOR == 4 && ActiveRecord::VERSION::MINOR == 2
-    end
-  end
-
-  module Error
-    class Standard < StandardError; end
-    class NotImplementedFeature < Standard
-      def message
-        "This Adapter doesn't offer updating single rows, Google Big query is append only by design"
-      end
-    end
-    class NotImplementedColumnOperation < Standard
-      def message
-        "Google big query doesn't allow this column operation"
-      end
-    end
-
-    class PendingFeature < Standard
-      def message
-        "Sorry, this is a pending feature, it will be implemented soon."
-      end
-    end
-  end
-
-  module ConnectionHandling # :nodoc:
-    # bigquery adapter reuses GoogleBigquery::Auth.
-    def bigquery_connection(config)
-
-      # Require database.
-      unless config[:database]
-        raise ArgumentError, "No database file specified. Missing argument: database"
-      end
-      db = GoogleBigquery::Auth.authorized? ? GoogleBigquery::Auth.client : GoogleBigquery::Auth.new.authorize
-      #db #quizas deberia ser auth.api o auth.client
-
-      #In case we are using a bigquery adapter as standard config in database.yml
-      #All models are BigQuery enabled
-      ActiveRecord::Base.send :include, ActiveRecord::BigQueryPersistence
-      ActiveRecord::SchemaMigration.send :include, ActiveRecord::BigQuerySchemaMigration
-      ActiveRecord::Migrator.send :include, ActiveRecord::BigQueryMigrator
-      ActiveRecord::Relation.send :include, ActiveRecord::BigQueryRelation
-      ActiveRecord::Base.send :include, ActiveRecord::BigQuerying
-
-      #db.busy_timeout(ConnectionAdapters::SQLite3Adapter.type_cast_config_to_integer(config[:timeout])) if config[:timeout]
-      ConnectionAdapters::BigqueryAdapter.new(db, logger, config)
-    rescue  => e
-      raise e
-      #Errno::ENOENT => error
-      #if error.message.include?("No such file or directory")
-      #  raise ActiveRecord::NoDatabaseError.new(error.message)
-      #else
-      #  raise error
-      #end
-    end
-  end
-
-  module BQConnector
-    extend ActiveSupport::Concern
-    module ClassMethods
-      def establish_bq_connection(path)
-        self.send :include, ActiveRecord::BigQueryPersistence
-        self.send :include, ActiveRecord::BigQueryRelation
-        self.send :include, ActiveRecord::BigQuerying
-        establish_connection path
-      end
-    end
-  end
-
-  ActiveRecord::Base.send :include, BQConnector
-
-
-  # = Active Record Persistence
-  module BigQueryPersistence
-    extend ActiveSupport::Concern
-
-    def delete
-      raise Error::NotImplementedFeature
-    end
-
-    module ClassMethods
-
-    end
-
-    private
-    # Creates a record with values matching those of the instance attributes
-    # and returns its id.
-    def create_record(attribute_names = @attributes.keys)
-      record_timestamps_hardcoded
-      attributes_values = self.changes.values.map(&:last)
-
-      row_hash = Hash[ [ self.changes.keys, attributes_values ].transpose ]
-      new_id =  SecureRandom.hex
-      @rows =   {"rows"=> [{
-                            "insertId"=> Time.now.to_i.to_s,
-                            "json"=> row_hash.merge("id"=> new_id)
-                          }]
-                }
-      conn_cfg = self.class.connection_config
-      result = GoogleBigquery::TableData.create(conn_cfg[:project],
-        conn_cfg[:database],
-        self.class.table_name ,
-        @rows )
-
-      #raise result["error"]["errors"].map{|o| "[#{o['domain']}]: #{o['reason']} #{o['message']}" }.join(", ") if result["error"].present?
-      #here we output the IN MEMORY id , because of the BQ latency
-      self.id = new_id #||= new_id if self.class.primary_key
-
-      @new_record = false
-      id
-    end
-
-    #Partially copied from activerecord::Timezones
-    def record_timestamps_hardcoded
-      if self.record_timestamps
-        current_time = current_time_from_proper_timezone
-
-        all_timestamp_attributes.each do |column|
-          if respond_to?(column) && respond_to?("#{column}=") && self.send(column).nil?
-            write_attribute(column.to_s, current_time)
-          end
-        end
-      end
-    end
-
-    # DISABLED FEATURE, Google Big query is append only by design.
-    def update_record(attribute_names = @attributes.keys)
-      raise Error::NotImplementedFeature
-    end
-  end
-
-  # = Active Record Quering
-  module BigQuerying
-    def find_by_sql(sql, binds = [])
-      cfg = ActiveRecord::Base.connection_config
-      result_set = connection.select_all(sanitize_sql(sql), "#{name} Load", binds)
-      column_types = {}
-
-      if result_set.respond_to? :column_types
-        column_types = result_set.column_types
-      else
-        ActiveSupport::Deprecation.warn "the object returned from `select_all` must respond to `column_types`"
-      end
-      # When AR BigQuery queries uses joins , the fields appear as [database.table].field ,
-      # so at least we clean the class columns to initialize the record propperly
-      #"whoa1393194159_users_id".gsub(/#{@config[:database]}_#{self.table_name}_/, "")
-      result_set.instance_variable_set("@columns", result_set.columns.map{|o| o.gsub(/#{cfg[:database]}_#{self.table_name}_/, "") } )
-
-      result_set.map { |record| instantiate(record, column_types) }
-    end
-  end
-
-  # = Active Record Relation
-  module BigQueryRelation
-    extend ActiveSupport::Concern
-      module ClassMethods
-        def delete(id_or_array)
-          raise Error::NotImplementedFeature
-        end
-
-        def update(id, attributes)
-          raise Error::NotImplementedFeature
-        end
-
-        def destroy_all(conditions = nil)
-          raise Error::NotImplementedFeature
-        end
-
-        def destroy(id)
-          raise Error::NotImplementedFeature
-        end
-
-        def delete_all(conditions = nil)
-          raise Error::NotImplementedFeature
-        end
-
-        def update_all(updates)
-          raise Error::NotImplementedFeature
-        end
-      end
-  end
-
-  module BigQuerySchemaMigration
-
-    def self.included base
-      attr_accessor :migration_file_pwd
-      base.instance_eval do
-        def schema_migration_hash
-          file = schema_migration_file("r")
-          json = JSON.parse(file.read)
-        end
-
-        def schema_migration_path
-          Dir.pwd + "/db/schema_migrations.json"
-        end
-
-        def schema_migration_file(mode="w+")
-          file_pwd = Dir.pwd + "/db/schema_migrations.json"
-          File.open( file_pwd, mode )
-        end
-
-        def create_table(limit=nil)
-          @migration_file_pwd = Dir.pwd + "/db/schema_migrations.json"
-          unless File.exists?(@migration_file_pwd)
-            puts "SCHEMA MIGRATION HERE"
-            version_options = {null: false}
-            version_options[:limit] = limit if limit
-
-            #connection.create_table(table_name, id: false) do |t|
-            #  t.column :version, :string, version_options
-            #end
-            file = schema_migration_file
-            file.puts({ db:{ table_name.to_sym => [] } }.to_json )
-            file.close
-            #connection.add_index table_name, :version, unique: true, name: index_name
-          end
-        end
-
-        #def self.drop_table
-        #  binding.pry
-        #  File.delete(schema_migration_path)
-        #end
-
-        def delete_version(options)
-          #versions = ActiveRecord::SchemaMigration.where(:version => version.to_s)
-          version = options[:version]
-          new_data = SchemaMigration.schema_migration_hash["db"]["schema_migrations"].delete_if{|o| o["version"] == version.to_s}
-          hsh = {:db=>{:schema_migrations => new_data } }
-          f = schema_migration_file
-          f.puts hsh.to_json
-          f.close
-        end
-
-        def create!(args, *opts)
-          current_data = schema_migration_hash
-          unless schema_migration_hash["db"]["schema_migrations"].map{|o| o["version"]}.include?(args[:version].to_s)
-            hsh = {:db=>{:schema_migrations => current_data["db"]["schema_migrations"] << args } }
-            f = schema_migration_file
-            f.puts hsh.to_json
-            f.close
-          end
-          true
-        end
-
-        def all
-          schema_migration_hash["db"]["schema_migrations"]
-        end
-
-        def where(args)
-          all.select{|o| o[args.keys.first.to_s] == args.values.first}
-        end
-      end
-    end
-  end
-
-  module BigQueryMigrator
-
-    def self.included base
-      #overload class methods
-      base.instance_eval do
-        def get_all_versions
-          SchemaMigration.all.map { |x| x["version"].to_i }.sort
-        end
-
-        def current_version
-          sm_table = schema_migrations_table_name
-          migration_file_pwd = Dir.pwd + "/db/schema_migrations.json"
-
-          if File.exists?(migration_file_pwd)
-            get_all_versions.max || 0
-          else
-            0
-          end
-        end
-
-        def needs_migration?
-          current_version < last_version
-        end
-
-        def last_version
-          get_all_versions.min.to_i
-          #last_migration.version
-        end
-
-        def last_migration #:nodoc:
-          migrations(migrations_paths).last || NullMigration.new
-        end
-
-      end
-      #overload instance methods
-      base.class_eval do
-        def current_version
-          migrated.max || 0
-        end
-
-        def current_migration
-          migrations.detect { |m| m["version"] == current_version }
-        end
-
-        #def migrated
-        #  @migrated_versions ||= Set.new(self.class.get_all_versions)
-        #end
-
-        private
-
-          def record_version_state_after_migrating(version)
-
-            if down?
-              migrated.delete(version)
-              ActiveRecord::SchemaMigration.delete_version(:version => version.to_s)
-            else
-              migrated << version
-              ActiveRecord::SchemaMigration.create!(:version => version.to_s)
-            end
-          end
-      end
-    end
-
-    #alias :current :current_migration
-  end
-
-  module LoadOperations
-    extend ActiveSupport::Concern
-    module ClassMethods
-      def bigquery_export(bucket_location = nil)
-        bucket_location = bucket_location.nil? ? "#{table_name}.json" : bucket_location
-        cfg = connection_config
-        GoogleBigquery::Jobs.export(cfg[:project],
-          cfg[:database],
-          table_name,
-          "#{cfg[:database]}/#{bucket_location}")
-      end
-
-      def bigquery_load(bucket_location = [])
-        bucket_location = bucket_location.empty? ? ["#{cfg[:database]}/#{table_name}.json"] : bucket_location
-        cfg = connection_config
-        fields = columns.map{|o| {name: o.name, type: o.sql_type, mode: "nullable" } }
-        GoogleBigquery::Jobs.load(cfg[:project],
-          cfg[:database],
-          table_name,
-          bucket_location,
-          fields)
-      end
-
-      def bigquery_import()
-      end
-    end
-  end
-
-  ActiveRecord::Base.send :include, LoadOperations
 
   module ConnectionAdapters
 
@@ -388,9 +26,45 @@ module ActiveRecord
       end
     end
 
+    class BQBinary < Type::Binary # :nodoc:
+      def cast_value(value)
+        if value.encoding != Encoding::ASCII_8BIT
+          value = value.force_encoding(Encoding::ASCII_8BIT)
+        end
+        value
+      end
+    end
+
+    class BQString < Type::String # :nodoc:
+      def type_cast_for_database(value)
+        binding.pry
+        if value.is_a?(::String) && value.encoding == Encoding::ASCII_8BIT
+          value.encode(Encoding::UTF_8)
+        else
+          super
+        end
+      end
+
+      def type_cast_from_user(value)
+        binding.pry
+      end
+    end
+
     class BigqueryAdapter < AbstractAdapter
 
-      include SchemaStatements
+      #include SchemaStatements
+
+      NATIVE_DATABASE_TYPES = {
+        :primary_key => "STRING",
+        :string      => { :name => "STRING", :default=> nil },
+        :integer     => { :name => "INTEGER", :default=> nil },
+        :float       => { :name => "FLOAT", :default=> 0.0 },
+        :datetime    => { :name => "TIMESTAMP" },
+        :timestamp    => { name: "TIMESTAMP" },
+        :date        => { :name => "TIMESTAMP" },
+        :record      => { :name => "RECORD" },
+        :boolean     => { :name => "BOOLEAN" }
+      }
 
       class Version
       end
@@ -464,6 +138,7 @@ module ActiveRecord
         end
       end
 
+
       class BindSubstitution < Arel::Visitors::SQLite # :nodoc:
         include Arel::Visitors::BindVisitor
       end
@@ -476,13 +151,14 @@ module ActiveRecord
                                         self.class.type_cast_config_to_integer(config.fetch(:statement_limit) { 1000 }))
         @config = config
         @config.merge(prepared_statements: false) if BigQueryRailsHelpers.rails40?
-
+        @prepared_statements = false
         #if self.class.type_cast_config_to_boolean(config.fetch(:prepared_statements) { true })
         #  @prepared_statements = true
         #  @visitor = Arel::Visitors::SQLite.new self
         #else
         #use the sql without prepraded statements, as I know BQ doesn't support them.
-
+        @type_map = Type::HashLookupTypeMap.new
+        initialize_type_map(type_map)
         @visitor = unprepared_visitor unless ActiveRecord::VERSION::MINOR >= 2
       end
 
@@ -576,21 +252,7 @@ module ActiveRecord
       end
 
       def native_database_types #:nodoc:
-        {
-          :primary_key => default_primary_key_type,
-          :string      => { :name => "STRING", :default=> nil },
-          #:text        => { :name => "text" },
-          :integer     => { :name => "INTEGER", :default=> nil },
-          :float       => { :name => "FLOAT", :default=> 0.0 },
-          #:decimal     => { :name => "decimal" },
-          :datetime    => { :name => "TIMESTAMP" },
-          #:timestamp   => { :name => "datetime" },
-          :timestamp    => { name: "TIMESTAMP" },
-          #:time        => { :name => "time" },
-          :date        => { :name => "TIMESTAMP" },
-          :record      => { :name => "RECORD" },
-          :boolean     => { :name => "BOOLEAN" }
-        }
+        NATIVE_DATABASE_TYPES
       end
 
       # Returns the current database encoding format as a string, eg: 'UTF-8'
@@ -623,7 +285,7 @@ module ActiveRecord
 
       # QUOTING ==================================================
 
-      def quote(value, column = nil)
+      def _quote(value, column = nil)
         if value.kind_of?(String) && column && column.type == :binary && column.class.respond_to?(:string_to_binary)
           s = column.class.string_to_binary(value).unpack("H*")[0]
           "x'#{s}'"
@@ -663,6 +325,7 @@ module ActiveRecord
       end
 
       def type_cast(value, column) # :nodoc:
+        binding.pry
         return value.to_f if BigDecimal === value
         return super unless String === value
         return super unless column && value
@@ -696,6 +359,7 @@ module ActiveRecord
       end
 
       def exec_query(sql, name = nil, binds = [])
+        binding.pry
         log(sql, name, binds) do
 
           # Don't cache statements if they are not prepared
@@ -729,6 +393,7 @@ module ActiveRecord
       end
 
       def insert_sql(sql, name = nil, pk = nil, id_value = nil, sequence_name = nil) #:nodoc:
+        binding.pry
         super
         id_value || @connection.last_insert_row_id
       end
@@ -787,16 +452,6 @@ module ActiveRecord
         #exec_query "DROP INDEX #{quote_column_name(index_name)}"
       end
 
-      def add_column(table_name, column_name, type, options = {}) #:nodoc:
-        if supports_add_column? && valid_alter_table_options( type, options )
-          super(table_name, column_name, type, options)
-        else
-          alter_table(table_name) do |definition|
-            definition.column(column_name, type, options)
-          end
-        end
-      end
-
       # See also TableDefinition#column for details on how to create columns.
       def create_table(table_name, options = {})
         td = create_table_definition table_name, options[:temporary], options[:options]
@@ -849,15 +504,14 @@ module ActiveRecord
 
       # See: http://www.sqlite.org/lang_altertable.html
       # SQLite has an additional restriction on the ALTER TABLE statement
-      def valid_alter_table_options( type, options)
+      def valid_alter_table_type?(type)
         type.to_sym != :primary_key
       end
 
       def add_column(table_name, column_name, type, options = {}) #:nodoc:
+        if valid_alter_table_type?(type)
 
-        if supports_add_column? && valid_alter_table_options( type, options )
-
-          hsh = table_name.classify.constantize.columns.map { |c|  {"name"=> c.name, "type"=> c.type }  }
+          hsh = table_name.classify.constantize.columns.map { |c|  {"name"=> c.name, "type"=> c.cast_type }  }
           hsh << {"name"=> column_name, :type=> type}
           fields = [ fields: hsh ]
 
@@ -868,7 +522,6 @@ module ActiveRecord
              "tableId"  => table_name },
              "schema"   => fields,
             "description"=> "added from migration"} )
-
         else
           bypass_feature
         end
@@ -925,6 +578,14 @@ module ActiveRecord
 
 
       protected
+
+        def initialize_type_map(m)
+          super
+          puts "INITILIZLIE TYPES MAP"
+          m.register_type(/binary/i, BQBinary.new)
+          register_class_with_limit m, %r(char)i, BQString
+        end
+
         def select(sql, name = nil, binds = []) #:nodoc:
           exec_query(sql, name, binds)
         end
